requirements.txt
PyYAML~=6.0.2
click~=8.1.7
